+++ Ship_Game/Universe/SolarBodies/UnexploredPlanetScreen.cs
@@ -39,7 +39,7 @@ namespace Ship_Game
                 titleRect.Width = 365;
             }
             TitleBar = new Menu2(titleRect);
-            TitlePos = new Vector2((float)(titleRect.X + titleRect.Width / 2) - Fonts.Laserian14.MeasureString(p.Name).X / 2f, (float)(titleRect.Y + titleRect.Height / 2 - Fonts.Laserian14.LineSpacing / 2));
+            TitlePos = new Vector2(titleRect.X + titleRect.Width / 2 - Fonts.Laserian14.MeasureString(p.Name).X / 2f, titleRect.Y + titleRect.Height / 2 - Fonts.Laserian14.LineSpacing / 2);
             Rectangle leftRect = new Rectangle(5, titleRect.Y + titleRect.Height + 5, titleRect.Width, 
                 ScreenHeight - (titleRect.Y + titleRect.Height) - (int)(0.4f * ScreenHeight));
             if (leftRect.Height < 350)
@@ -56,17 +56,17 @@ namespace Ship_Game
         public override void Draw(SpriteBatch batch)
         {
             TitleBar.Draw();
-            float x = (float)Mouse.GetState().X;
+            float x = Mouse.GetState().X;
             MouseState state = Mouse.GetState();
-            Vector2 MousePos = new Vector2(x, (float)state.Y);
+            Vector2 MousePos = new Vector2(x, state.Y);
             Color c = new Color(255, 239, 208);
             batch.DrawString(Fonts.Laserian14, p.Name, TitlePos, c);
             PlanetMenu.Draw();
             PlanetInfo.Draw();
             batch.Draw(ResourceManager.Texture(string.Concat("Planets/", p.PlanetType)), PlanetIcon, Color.White);
-            Vector2 PNameCursor = new Vector2((float)(PlanetInfo.Menu.X + 20), (float)(PlanetInfo.Menu.Y + 45));
+            Vector2 PNameCursor = new Vector2(PlanetInfo.Menu.X + 20, PlanetInfo.Menu.Y + 45);
             batch.DrawString(Fonts.Arial20Bold, p.Name, PNameCursor, new Color(255, 239, 208));
-            PNameCursor.Y = PNameCursor.Y + (float)(Fonts.Arial20Bold.LineSpacing * 2);
+            PNameCursor.Y = PNameCursor.Y + Fonts.Arial20Bold.LineSpacing * 2;
             float amount = 80f;
             if (GlobalStats.IsGerman)
             {
@@ -78,13 +78,13 @@ namespace Ship_Game
             batch.DrawString(Fonts.Arial12Bold, p.GetTypeTranslation(), InfoCursor, new Color(255, 239, 208));
             if (!p.Habitable)
             {
-                PNameCursor.Y = PNameCursor.Y + (float)(Fonts.Arial12Bold.LineSpacing + 2);
+                PNameCursor.Y = PNameCursor.Y + (Fonts.Arial12Bold.LineSpacing + 2);
                 InfoCursor = new Vector2(PNameCursor.X + 80f, PNameCursor.Y);
                 batch.DrawString(Fonts.Arial12Bold, "Uninhabitable", PNameCursor, Color.Orange);
             }
             else
             {
-                PNameCursor.Y = PNameCursor.Y + (float)(Fonts.Arial12Bold.LineSpacing + 2);
+                PNameCursor.Y = PNameCursor.Y + (Fonts.Arial12Bold.LineSpacing + 2);
                 InfoCursor = new Vector2(PNameCursor.X + amount, PNameCursor.Y);
                 ScreenManager.SpriteBatch.DrawString(Fonts.Arial12Bold, string.Concat(Localizer.Token(385), ":"), PNameCursor, Color.Orange);
                 SpriteBatch spriteBatch1 = ScreenManager.SpriteBatch;
@@ -98,7 +98,7 @@ namespace Ship_Game
                 {
                     ToolTip.CreateTooltip(75);
                 }
-                PNameCursor.Y = PNameCursor.Y + (float)(Fonts.Arial12Bold.LineSpacing + 2);
+                PNameCursor.Y = PNameCursor.Y + (Fonts.Arial12Bold.LineSpacing + 2);
                 InfoCursor = new Vector2(PNameCursor.X + amount, PNameCursor.Y);
                 ScreenManager.SpriteBatch.DrawString(Fonts.Arial12Bold, string.Concat(Localizer.Token(386), ":"), PNameCursor, Color.Orange);
                 ScreenManager.SpriteBatch.DrawString(Fonts.Arial12Bold, p.Fertility.ToString(fmt), InfoCursor, new Color(255, 239, 208));
@@ -107,7 +107,7 @@ namespace Ship_Game
                 {
                     ToolTip.CreateTooltip(20);
                 }
-                PNameCursor.Y = PNameCursor.Y + (float)(Fonts.Arial12Bold.LineSpacing + 2);
+                PNameCursor.Y = PNameCursor.Y + (Fonts.Arial12Bold.LineSpacing + 2);
                 InfoCursor = new Vector2(PNameCursor.X + amount, PNameCursor.Y);
                 ScreenManager.SpriteBatch.DrawString(Fonts.Arial12Bold, string.Concat(Localizer.Token(387), ":"), PNameCursor, Color.Orange);
                 ScreenManager.SpriteBatch.DrawString(Fonts.Arial12Bold, p.MineralRichness.ToString(fmt), InfoCursor, new Color(255, 239, 208));
@@ -117,8 +117,8 @@ namespace Ship_Game
                     ToolTip.CreateTooltip(21);
                 }
             }
-            PNameCursor.Y = (float)(PlanetIcon.Y + PlanetIcon.Height + 20);
-            string desc = parseText(p.Description, (float)(PlanetInfo.Menu.Width - 40));
+            PNameCursor.Y = PlanetIcon.Y + PlanetIcon.Height + 20;
+            string desc = parseText(p.Description, PlanetInfo.Menu.Width - 40);
             batch.DrawString(Fonts.Arial12Bold, desc, PNameCursor, new Color(255, 239, 208));
             /*if (this.p.Special != "None")     //This was removed, because the string "Special" was never assigned a valus other than "None" -Gretman
             {
@@ -173,7 +173,7 @@ namespace Ship_Game
             string line = string.Empty;
             string returnString = string.Empty;
             string[] strArrays = text.Split(' ');
-            for (int i = 0; i < (int)strArrays.Length; i++)
+            for (int i = 0; i < strArrays.Length; i++)
             {
                 string word = strArrays[i];
                 if (Fonts.Arial12Bold.MeasureString(string.Concat(line, word)).Length() > Width)
